check existence of log files

+++ app/services/DailyReportService.php
@@ -347,9 +347,8 @@ class DailyReportService extends Injectable
     {
         $filename = BASE_DIR . '/app/logs/report.log';
 
-        if (filesize($filename) > 128*1024) {
+        if (file_exists($filename) && filesize($filename) > 128*1024) {
             unlink($filename);
-            touch($filename);
         }
 
         $str = date('Y-m-d H:i:s ') . $str . "\n";
